% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NfpmViz-class.R
\docType{class}
\name{NfpmViz-class}
\alias{NfpmViz-class}
\alias{NfpmViz}
\title{Data container for an nfpm Track.}
\description{
Data container for an nfpm Track.
}
\section{Methods}{

\describe{
\item{\code{get_attributes()}}{Get attributes for rendering component}

\item{\code{get_colors()}}{Get chart colors}

\item{\code{get_component_type()}}{Get component type for prefix of random id generator}

\item{\code{get_data()}}{Get chart data}

\item{\code{get_default_colors()}}{Get default colors}

\item{\code{get_default_settings()}}{Get default settings}

\item{\code{get_name()}}{Get name of Epiviz Web Component}

\item{\code{get_settings()}}{Get chart settings}

\item{\code{render_component(shiny = FALSE)}}{Render to html}

\item{\code{set_colors(colors)}}{Set chart colors}

\item{\code{set_data(data)}}{Set chart data}

\item{\code{set_settings(settings)}}{Modify current settings
\describe{
  \item{settings}{List of new settings.
    Call get_available_settings for settings available to modify.
  }
}}
}}

